+++ src/Symfony/Bundle/FrameworkBundle/Tests/Functional/KernelTestCaseTest.php
@@ -18,36 +18,27 @@ use Symfony\Bundle\FrameworkBundle\Tests\Functional\Bundle\TestBundle\TestServic
 use Symfony\Bundle\FrameworkBundle\Tests\Functional\Bundle\TestBundle\TestServiceContainer\UnusedPrivateService;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
-class TestServiceContainerTest extends AbstractWebTestCase
+class KernelTestCaseTest extends AbstractWebTestCase
 {
-    /**
-     * @group legacy
-     */
     public function testThatPrivateServicesAreUnavailableIfTestConfigIsDisabled()
     {
         static::bootKernel(['test_case' => 'TestServiceContainer', 'root_config' => 'test_disabled.yml', 'environment' => 'test_disabled']);
 
-        $this->assertInstanceOf(ContainerInterface::class, static::$container);
-        $this->assertNotInstanceOf(TestContainer::class, static::$container);
-        $this->assertTrue(static::$container->has(PublicService::class));
-        $this->assertFalse(static::$container->has(NonPublicService::class));
-        $this->assertFalse(static::$container->has(PrivateService::class));
-        $this->assertFalse(static::$container->has('private_service'));
-        $this->assertFalse(static::$container->has(UnusedPrivateService::class));
+        $this->expectException(\LogicException::class);
+        static::getContainer();
     }
 
-    /**
-     * @group legacy
-     */
     public function testThatPrivateServicesAreAvailableIfTestConfigIsEnabled()
     {
         static::bootKernel(['test_case' => 'TestServiceContainer']);
 
-        $this->assertInstanceOf(TestContainer::class, static::$container);
-        $this->assertTrue(static::$container->has(PublicService::class));
-        $this->assertTrue(static::$container->has(NonPublicService::class));
-        $this->assertTrue(static::$container->has(PrivateService::class));
-        $this->assertTrue(static::$container->has('private_service'));
-        $this->assertFalse(static::$container->has(UnusedPrivateService::class));
+        $container = static::getContainer();
+        $this->assertInstanceOf(ContainerInterface::class, $container);
+        $this->assertInstanceOf(TestContainer::class, $container);
+        $this->assertTrue($container->has(PublicService::class));
+        $this->assertTrue($container->has(NonPublicService::class));
+        $this->assertTrue($container->has(PrivateService::class));
+        $this->assertTrue($container->has('private_service'));
+        $this->assertFalse($container->has(UnusedPrivateService::class));
     }
 }