+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -16,6 +16,7 @@ CHANGELOG
  * Add tag `assets.package` to register asset packages
  * Add support to use a PSR-6 compatible cache for Doctrine annotations
  * Deprecate all other values than "none", "php_array" and "file" for `framework.annotation.cache`
+ * Add `KernelTestCase::getContainer()` as the best way to get a container in tests
 
 5.2.0
 -----